// front/src/store/action-types.ts
export enum ActionTypes {
  AIRLOCK_OPEN = 'AIRLOCK_OPEN',
  FLOCK_FORM = 'FLOCK_FORM',
  FLOCK_DROP = 'FLOCK_DROP',
  FLOCK_JOIN = 'FLOCK_JOIN',

  INVITE_SEND = 'INVITE_SEND',
  INVITE_RESCIND = 'INVITE_RESCIND',

  SCRY_AVAIL_CHATS = 'SCRY_AVAIL_CHATS',
  SCRY_FLOKS = 'SCRY_FLOKS',
  SCRY_SENT_INVITES = 'SCRY_SENT_INVITES',
  SCRY_PENDING_INVITES = 'SCRY_PENDING_INVITES',
}