// front/src/store/mutation-types.ts
export enum MutationTypes {
  AVAILABLE_CHATS_SET = 'AVAILABLE_CHATS_SET',

  EXISTING_FLOCKS_SET = 'EXISTING_FLOCKS_SET',
  EXISTING_FLOCKS_ADD = 'EXISTING_FLOCKS_ADD',
  EXISTING_FLOCKS_REMOVE = 'EXISTING_FLOCKS_REMOVE',

  SENT_INVITES_SET = 'SENT_INVITES_SET',
  SENT_INVITES_ADD = 'SENT_INVITES_ADD',

  PENDING_INVITES_SET = 'PENDING_INVITES_SET',
}